+++ src/view/CategorieRecette/list.php
@@ -2,6 +2,10 @@
 
 $object = static::$object;
 $primary = 'idCategorieRecette';
+
+$isConnected = Session::isConnected();
+$isAdmin = Session::isAdmin();
+
 $isUpdate = false;
 $idToUpdate = '';
 if ($_GET['action'] == 'readAll' && isset($_GET[$primary])) {
@@ -9,17 +13,18 @@ if ($_GET['action'] == 'readAll' && isset($_GET[$primary])) {
     $idToUpdate = $_GET[$primary];
 }
 
-echo <<< EOT
-    <div id="divCreation{$object}"> 
-        <form method="post" action="index.php?controller={$object}&action=created">
-            <label for="nomCategorieRecette" >Ajouter une catégorie : </label>
-            <input type="text" name="nomCategorieRecette">
-            <input type="hidden" name="controller" value="{$object}"/>
-            <input type="submit" value="Envoyer"/>
-        </form>
-    </div>
-        
+if ($isConnected) {
+    echo <<< EOT
+        <div id="divCreation{$object}"> 
+            <form method="post" action="index.php?controller={$object}&action=created">
+                <label for="nomCategorieRecette" >Ajouter une catégorie : </label>
+                <input type="text" name="nomCategorieRecette">
+                <input type="hidden" name="controller" value="{$object}"/>
+                <input type="submit" value="Envoyer"/>
+            </form>
+        </div>    
 EOT;
+}
 
 echo <<< EOT
     <div id="divList{$object}"> 
@@ -41,7 +46,7 @@ foreach ($tab_categorieRecette as $categorieRecette) {
                     
                     <input hidden name="{$primary}" value="{$spe_idCategorieRecette}">
                     <input type="hidden" name="controller" value="<?=static::$object?>"/>
-                    <button type="submit">Ajouter</button>
+                    <button type="submit">Valider</button>
               
                     <a href="./index.php?controller={$object}&action=delete&{$primary}={$raw_idCategorieRecette}">
                         <button type="button">Supprimer</button>
@@ -51,19 +56,28 @@ foreach ($tab_categorieRecette as $categorieRecette) {
 EOT;
     }
     else {
-
         echo <<< EOT
             <li>
                {$spe_nomCategorieRecette}
+EOT;
+
+        if ($isConnected) {
+            echo <<< EOT
                 <a href="./index.php?controller={$object}&action=readAll&{$primary}={$raw_idCategorieRecette}">
                     <button type="button">Modifier</button>
                 </a>
+EOT;
+        }
+
+        if ($isAdmin) {
+            echo <<< EOT
                 <a href="./index.php?controller={$object}&action=delete&{$primary}={$raw_idCategorieRecette}">
                     <button type="button">Supprimer</button>
                 </a>
-            </li>
 EOT;
+        }
 
+        echo '</li>';
     }
 }
